Finally fully functional.
TODO: add a 'refresh' when switching autopilot mode in flight (currently you have to go back to the Space Center and then back to the vessel for the change to fully propagate).

There also seems to be an issue with the crew bonus not being applied, afflicts both RoverController and ShipController.  Need to investigate further.